exam-backend/requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.18

# Database — SQLite is built-in (default); psycopg2 only if using PostgreSQL
sqlalchemy>=2.0.0
alembic>=1.14.0

# Celery + Redis (optional — falls back to sync mode if Redis not running)
celery[redis]>=5.4.0
redis>=5.0.0

# PDF + OCR
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0

# Google Gemini AI
google-generativeai>=0.7.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=21.0.0
tenacity>=8.0.0
